Make generated arrays available at compile time
A bug in earlier perls is being fixed in v5.30.  See
https://rt.perl.org/Ticket/Display.html?id=133860

This bug allowed this module to work, by luck.  But now it doesn't.  The
ticket whose link is given above has details, but in short, the variable
@IGNORED needs to be populated at compile time.  This patch changes so
the loop that generates the code causes all similar variables are
enclosed in BEGIN blocks.

Since this module is important to CPAN, with many other modules
dependent upon it, it would be good to get this fixed as soon as
possible.

There may be better solutions, but I found this without having to know
much about the workings of the module, and it seems safe to me.